Change TLS 1.2 SSL Factory Creation to be based On the Platform
The sdk takes the platform and tries to understand whether the TLS installation is installed but not
enabled. If this is the case, it recommends the usage of the SDK provided implementation of the ssl factory
This behaviour can be overwritten by using the new `.setTls12Implementation` on the client builder. This
can be set to always be using the sdk implementation, the system, or the default recommendation.

It is recommended to leave the recommendation to use the system provided SSL factory, so to not interfere
with other already setup settings. Only if the system does not enable TLS 1.2 by default, the custom
factory is used.