feat: Implement Serializable interface for rich text classes
The changes in this commit add the Serializable interface to various classes in the `com.contentful.java.cda.rich` package. This allows these classes to be easily serialized and deserialized, which is important for use cases where the rich text data needs to be persisted or transmitted over a network.

The main classes that now implement Serializable are:

- `CDARichTableCell`
- `CDARichQuote`
- `CDARichListItem`
- `CDARichTableRow`
- `CDARichTable`
- `CDARichText`
- `CDARichNode`
- `CDARichMark`
- `CDARichHorizontalRule`
- `CDARichList`
- `CDARichUnorderedList`
- `CDARichEmbeddedInline`
- `CDARichHyperLink`

This change will enable better integration with other systems and frameworks that require serializable data, and will improve the overall flexibility and usability of the rich text handling functionality in the Contentful Java CDA SDK.